Implement HAPROXY_STICKY_COOKIE in haproxy_template.cfg
If the HAPROXY_STICKY_COOKIE environment variable is defined in a
Marathon application registered to Bamboo, Bamboo will implement sticky
sessions for that application according to the value of the variable.

For example, if HAPROXY_STICKY_COOKIE="SERVERID insert", then Bamboo
will set a cookie named SERVERID to the value of internal IP and port of
the task instance to stick the user session to.